refactor: modularized daser and simplified node/module.go
